refactor: extract semantic tokens logic to handler module
- Create src/handlers/semantic_tokens.rs with semantic tokens logic
- Move LEGEND_TYPES constant to the handler module where it belongs
- Simplify lib.rs by delegating to handler function
- Use underscore separator in filename following Rust conventions

This continues the modularization effort, separating LSP protocol
handling from implementation details.
