+++ src/handlers/mod.rs
@@ -1,5 +1,7 @@
 pub mod definition;
+pub mod semantic_tokens;
 
 pub use definition::{
     ContextType, DefinitionCandidate, DefinitionResolver, ReferenceContext, handle_goto_definition,
 };
+pub use semantic_tokens::{LEGEND_TYPES, handle_semantic_tokens_full};
